move linear histogram matching to utils.py
