+++ .gitignore
@@ -1,0 +1,5 @@
+#Ignore __pycache__
+/__pycache__/
+
+#Ignore results
+/results/
